updated file exercises/078acc7a-bfc0-48df-87b8-4a2ce38c252e/solutions/28acd6f8-4e17-47c7-8068-29d141a6ebbb/sol.py

+++ exercises/078acc7a-bfc0-48df-87b8-4a2ce38c252e/solutions/28acd6f8-4e17-47c7-8068-29d141a6ebbb/sol.py
@@ -1,7 +1,7 @@
 print('Number of children?')
 ile = int(input())
 print('Income per member?')
-income_percapita = int(input())
+income_percapita = float(input())
 plus = 200
 if income_percapita <= 500:
     print(f'Total children benefit: {plus*(ile)}')
